# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Task Management backend server",
        "title": "heisenflow-System",
        "contact": {
            "name": "HeisenGo",
            "url": "https://github.com/HeisenGo"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/columns": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reorder the columns of a board for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Columns"
                ],
                "summary": "Reorder columns",
                "parameters": [
                    {
                        "description": "Reorder Columns Request",
                        "name": "ReorderColumnsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.ReorderColumnsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Columns reordered successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.ColumnResponseItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid reorder details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden, permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create multiple columns for a specified board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Columns"
                ],
                "summary": "Create columns",
                "parameters": [
                    {
                        "description": "Columns creation details",
                        "name": "columns",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.CreateColumnsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "response: details of created columns",
                        "schema": {
                            "$ref": "#/definitions/presenter.CreateColumnsResponse"
                        }
                    },
                    "400": {
                        "description": "error: bad request, invalid board ID format or missing columns details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/columns/{columnID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a column by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Columns"
                ],
                "summary": "Delete column",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Column ID",
                        "name": "columnID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "error: bad request, invalid column ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error: not found, column not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "presenter.ColumnResponseItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.CreateColumnItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "presenter.CreateColumnsRequest": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.CreateColumnItem"
                    }
                }
            }
        },
        "presenter.CreateColumnsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ColumnResponseItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "presenter.ReorderColumnItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "presenter.ReorderColumnsRequest": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ReorderColumnItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type 'Bearer' followed by a space and your JWT token.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}